backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
bcrypt==4.1.2
black==25.11.0
boto3==1.40.76
botocore==1.40.76
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.3
flake8==7.3.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
jq==1.10.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.1.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.6.1
pytest==9.0.1
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytokens==0.3.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.50.0
tenacity==8.2.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
yarl==1.22.0
openai==1.52.0
xlsxwriter==3.1.9
